chore(agentic-refactor): rename missed package from ai to agentic
