+++ rover/internal/handler/rover/agentic/exposure.go
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-package ai
+package agentic
 
 import (
 	"context"
