+++ rover/internal/handler/rover/agentic/subscription.go
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-package ai
+package agentic
 
 import (
 	"context"
